tests(helpers): separate customized asserts (#13628)
KAG-5252
